requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.5
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-phonenumber-field==6.1.0
django-rest-swagger==2.2.0
django-seed==0.3.1
django-smtp-ssl==1.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
Faker==13.3.2
idna==3.3
importlib-metadata==4.8.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.2
kombu==5.1.0
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.0
phonenumbers==8.12.44
Pillow==9.0.1
pluggy==1.0.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
requests==2.26.0
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
toposort==1.7
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
zipp==3.6.0
